Add files via upload
Please attach the TerrainGen to the background. Currently, it only spawns pieces with the tag "terrain" so change/add that tag to your prefabs. Once you attach the script to the background and (hopefully) look through the code, you'll notice that the array which holds the terrain is public. I've left it public for now so that we can easily test out how the terrain pieces look. I've also left speed public so everyone can test out and agree on what would be a good starting speed.